# swagger.json
{
  "openapi": "3.0.0",
  "info": {
  "title": "API de Transferências e Usuários",
    "version": "1.0.0",
    "description": "API para registro, login, consulta de usuários e transferências."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "inserirSaldo": { "type": "number", "minimum": 100 },
                  "favorecidos": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["username", "password", "inserirSaldo"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado" },
          "400": { "description": "Usuário já existe" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "401": { "description": "Login ou senha inválidos" }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Listar usuários",
        "responses": {
          "200": { "description": "Lista de usuários" }
        }
      }
    },
    "/transfer": {
      "post": {
        "summary": "Realizar transferência",
        "security": [ { "bearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": { "type": "string" },
                  "to": { "type": "string" },
                  "amount": { "type": "number" }
                },
                "required": ["from", "to", "amount"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transferência realizada" },
          "400": { "description": "Erro na transferência" },
          "401": { "description": "Token não fornecido ou inválido" },
          "403": { "description": "Token inválido" }
        }
      }
    },
    "/transfers": {
      "get": {
        "summary": "Listar transferências",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "Lista de transferências" },
          "401": { "description": "Token não fornecido ou inválido" },
          "403": { "description": "Token inválido" }
        }
      }
    }
  }
}